# factory-saas/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 厂区访客管控 SaaS API
  description: |
    Factory Visitor Control SaaS — 多租户访客手机管控系统 API。
    
    ## 认证方式
    - Cookie 认证：登录后通过 `token` HttpOnly cookie 进行身份验证
    - CSRF 保护：非安全方法（POST/PUT/DELETE）需携带 `X-CSRF-Token` 头，值与 `csrf_token` cookie 一致
    
    ## 错误格式
    所有 API 错误返回统一格式：
    ```json
    {
      "error": "ERROR_CODE",
      "message": "用户可见描述",
      "timestamp": "2024-01-01T00:00:00.000Z",
      "requestId": "uuid"
    }
    ```
  version: 2.0.0
  contact:
    name: Factory SaaS Support
servers:
  - url: http://localhost:3000
    description: 开发环境
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subscriptionId:
          type: string
        visitorName:
          type: string
        visitorCompany:
          type: string
        area:
          type: string
        status:
          type: string
          enum: [waiting, pairing, paired_not_connected, restricted, exiting, exited, error]
        deviceId:
          type: string
          nullable: true
        deviceIp:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        restrictedAt:
          type: string
          format: date-time
          nullable: true
        exitedAt:
          type: string
          format: date-time
          nullable: true
        logs:
          type: array
          items:
            type: string
        entryQR:
          type: string
          description: Base64 data URL
        exitQR:
          type: string
          description: Base64 data URL
        selfCheckin:
          type: boolean
        tamperDetected:
          type: boolean
...
  parameters:
    csrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
paths:
...
  /api/sessions:
    get:
      tags: [sessions]
      summary: 获取当前活跃会话列表
      security:
        - cookieAuth: []
      parameters:
        - name: siteId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
  /api/sessions/{id}/exit:
    post:
      tags: [sessions]
      summary: 访客正常离厂
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/csrfHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                exitToken:
                  type: string
      responses:
        '200':
          description: 离厂成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  tamperDetected:
                    type: boolean
                  verifyReport:
                    type: object
  /api/sessions/{id}/force-exit:
    post:
      tags: [sessions]
      summary: 强制离厂（操作员）
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/csrfHeader'
      responses:
        '200':
          description: 强制离厂成功
  /api/sessions/{id}/retry:
    post:
      tags: [sessions]
      summary: 重试管控
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/csrfHeader'
      responses:
        '200':
          description: 重试成功
  /api/sessions/{id}/device:
    post:
      tags: [sessions]
      summary: 上报设备 IP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/csrfHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [deviceIp]
              properties:
                deviceIp:
                  type: string
                  format: ipv4
      responses:
        '200':
          description: 设备 IP 已记录

  # ── History ─────────────────────────────────────────────────
  /api/history/sessions:
    get:
      tags: [sessions]
      summary: 查询会话历史
      security:
        - cookieAuth: []
      parameters:
        - name: siteId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            default: exited
        - name: q
          in: query
          schema:
            type: string
          description: 关键词搜索
        - name: from
          in: query
          schema:
            type: integer
          description: 起始时间戳
        - name: to
          in: query
          schema:
            type: integer
          description: 结束时间戳
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 500
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  total:
                    type: integer
                  hasMore:
                    type: boolean
  /api/history/sessions/export:
    get:
      tags: [sessions]
      summary: 导出会话历史 CSV
      security:
        - cookieAuth: []
      parameters:
        - name: siteId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: CSV 文件
          content:
            text/csv:
              schema:
                type: string

  # ── Sites ───────────────────────────────────────────────────
...